% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_full_rank_matrix.R
\name{make_full_rank_matrix}
\alias{make_full_rank_matrix}
\title{Create a full rank matrix. The function first removes empty columns. The it discovers linear dependent columns, then renames and removes columns to make the matrix full rank.}
\usage{
make_full_rank_matrix(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A matrix of full rank.
}
\description{
Create a full rank matrix. The function first removes empty columns. The it discovers linear dependent columns, then renames and removes columns to make the matrix full rank.
}
\examples{
c1 <- rbinom(10, 1, .4)
c2 <- 1-c1
c3 <- integer(10)
c4 <- c1
mat <- as.matrix(data.frame(c1, c2, c3, c4))
make_full_rank_matrix(mat)
}
